スケジュールに関する部分を削除,整備情報のコンポーネント削除、
