--- src/components/pages/Schedule.tsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-export const Schedule: React.FC =() => {
-  return(
-    <>
-      <div>スケジュールページ</div>
-    </>
-  )
-}
