+++ src/components/utils/MenuList.tsx
@@ -2,7 +2,6 @@ import React from 'react';
 
 import { ListItemIcon, ListItemButton, ListItemText, Typography}  from '@mui/material';
 import AirplanemodeActiveIcon from '@mui/icons-material/AirplanemodeActive';
-import CalendarMonthIcon from '@mui/icons-material/CalendarMonth';
 import PeopleAltIcon from '@mui/icons-material/PeopleAlt';
 
 import { Link } from "react-router-dom"
@@ -15,14 +14,6 @@ export const MenuList = (
           < AirplanemodeActiveIcon />
         </ListItemIcon>
         <ListItemText primary="機材情報" />
-      </ListItemButton>
-    </Typography>
-    <Typography component={Link} sx={{textDecoration: 'none', color: 'inherit'}} to="/schedule">
-      <ListItemButton>
-        <ListItemIcon>
-          < CalendarMonthIcon />
-        </ListItemIcon>
-        <ListItemText primary="スケジュール情報" />
       </ListItemButton>
     </Typography>
     <ListItemButton>
