+++ src/lib/api/ship.ts
@@ -4,10 +4,6 @@ export const ship = () => {
   return client.get('/ships');
 };
 
-export const schedule = (id: string | undefined) => {
-  return client.get(`ships/${id}/schedules`)
-}
-
 export const maintenance = (id: string | undefined) => {
   return client.get(`ships/${id}/maintenances`)
 }
