--- src/types/schedule.ts
@@ -1,10 +1,0 @@
-export type ScheduleData = {
-  id: number
-  flightNumber: string
-  from: string
-  to: string
-  depDate: string
-  depTime: string
-  arrDate: string
-  arrTime: string
-}
